+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Table of Contents
 
+* [2.5.0 - 2024-10-21](#250---2024-10-21)
 * [2.4.1 - 2024-10-13](#241---2024-10-13)
 * [2.4.0 - 2024-09-21](#240---2024-09-21)
 * [2.3.0 - 2024-03-17](#230---2024-03-17)
@@ -21,6 +22,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [1.1.0 - 2020-05-08](#110---2020-05-08)
 * [1.0.0 - 2020-05-05](#100---2020-05-05)
 * [0.0.0 - 2020-03-17](#000---2020-03-17)
+
+## 2.5.0 - 2024-10-21
+
+* Added
+  * Development time functionality to export specs as JSON Schema
+* Fixed
+  * Updated the sample data in each spec to render as the correct type
 
 ## 2.4.1 - 2024-10-13